# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Go License API
  version: 1.0.0
  description: |
    High-performance multi-tenant license validation service.
    Data Plane under `/licenses/*`, Control Plane under `/admin/*`.
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /admin/tenants/{id}:
    delete:
      summary: Delete tenant
      operationId: deleteTenant
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    TenantAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security: []